Update prduction (#5)
* update (docs) update intents docs & engines docs

* update (routes) add filter engine by brand, update models, refactor payment itents

* update (route) update cars route add a deletion of owned cars etc.

* add (models) update car model with photos

* add (content) add privacy policy for app

* add (cars routes) add functionality for updating car

* update (routes) update routes (add some for cars  add features for uploading multy photos start to do router for details sale)

* update (docs) add some docs for routing of cars

+++ models/payment_intent.go
@@ -25,6 +25,6 @@ type PaymentIntent struct {
 	Currency     string  `json:"currency" gorm:"not null"`
 	CustomerID   string  `json:"customer" gorm:"not null"`
 	CanceledAt   int64   `json:"canceled" gorm:"not null"`
-	Amount       float32 `json:"amount" gorm:"not null"`
+	Amount       int64 `json:"amount" gorm:"not null"`
 	ClientSecret string  `json:"secret" gorm:"not null"`
 }
